Refactored the GUI
